% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colVars.R
\name{colVars}
\alias{colVars}
\title{FUNCTION_TITLE}
\usage{
colVars(a)
}
\arguments{
\item{a}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
