% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileModel.R
\name{compileModel}
\alias{compileModel}
\title{FUNCTION_TITLE}
\usage{
compileModel(model, stanDir = stanDir)
}
\arguments{
\item{model}{PARAM_DESCRIPTION}

\item{stanDir}{PARAM_DESCRIPTION, Default: stanDir}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
