% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimsTable.R
\name{getSimsTable}
\alias{getSimsTable}
\title{FUNCTION_TITLE}
\usage{
getSimsTable(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}
