% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit.inv}
\alias{logit.inv}
\alias{logit}
\title{generalized logit and inverse logit functions}
\usage{
logit.inv(x, lower = 0, upper = 1)

logit(x, lower = 0, upper = 1)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{lower}{PARAM_DESCRIPTION, Default: 0}

\item{upper}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
