% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mbmaPpcPlot}
\alias{mbmaPpcPlot}
\title{FUNCTION_TITLE}
\usage{
mbmaPpcPlot(data, x = "x", treatment = "treatment", type = "type",
  xlab = "x")
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION, Default: 'x'}

\item{treatment}{PARAM_DESCRIPTION, Default: 'treatment'}

\item{type}{PARAM_DESCRIPTION, Default: 'type'}

\item{xlab}{PARAM_DESCRIPTION, Default: 'x'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lattice]{panel.xyplot}},\code{\link[lattice]{panel.segments}}
}
