% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcDensityOld.R
\name{mcmcDensityOld}
\alias{mcmcDensityOld}
\title{FUNCTION_TITLE}
\usage{
mcmcDensityOld(fit, pars = names(fit), byChain = FALSE, nParPerPage = 16,
  prior = NULL)
}
\arguments{
\item{fit}{PARAM_DESCRIPTION}

\item{pars}{PARAM_DESCRIPTION, Default: names(fit)}

\item{byChain}{PARAM_DESCRIPTION, Default: FALSE}

\item{nParPerPage}{PARAM_DESCRIPTION, Default: 16}

\item{prior}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
