% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcHistory.R
\name{mcmcHistory}
\alias{mcmcHistory}
\title{FUNCTION_TITLE}
\usage{
mcmcHistory(fit, pars = names(fit), nParPerPage = 6, myTheme = NULL)
}
\arguments{
\item{fit}{PARAM_DESCRIPTION}

\item{pars}{PARAM_DESCRIPTION, Default: names(fit)}

\item{nParPerPage}{PARAM_DESCRIPTION, Default: 6}

\item{myTheme}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[bayesplot]{nuts_params}},\code{\link[bayesplot]{mcmc_trace}}
}
