% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcHistoryOld.R
\name{mcmcHistoryOld}
\alias{mcmcHistoryOld}
\title{FUNCTION_TITLE}
\usage{
mcmcHistoryOld(fit, pars = names(fit), nParPerPage = 6)
}
\arguments{
\item{fit}{PARAM_DESCRIPTION}

\item{pars}{PARAM_DESCRIPTION, Default: names(fit)}

\item{nParPerPage}{PARAM_DESCRIPTION, Default: 6}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
