% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mvdnorm}
\alias{mvdnorm}
\title{FUNCTION_TITLE}
\usage{
mvdnorm(y, mu, sigma2)
}
\arguments{
\item{y}{PARAM_DESCRIPTION}

\item{mu}{PARAM_DESCRIPTION}

\item{sigma2}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
