% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepanel.ci}
\alias{prepanel.ci}
\alias{panel.ci}
\title{FUNCTION_TITLE}
\usage{
prepanel.ci(y, ly, uy, subscripts, ...)

panel.ci(x, y, ly, uy, subscripts, pch = 16, col.line = "black", ...)
}
\arguments{
\item{y}{PARAM_DESCRIPTION}

\item{ly}{PARAM_DESCRIPTION}

\item{uy}{PARAM_DESCRIPTION}

\item{subscripts}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{pch}{PARAM_DESCRIPTION, Default: 16}

\item{col.line}{PARAM_DESCRIPTION, Default: 'black'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lattice]{panel.arrows}},\code{\link[lattice]{panel.xyplot}}
}
