% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterTable.R
\name{parameterTable}
\alias{parameterTable}
\title{FUNCTION_TITLE}
\usage{
parameterTable(fit, pars = names(fit))
}
\arguments{
\item{fit}{PARAM_DESCRIPTION}

\item{pars}{PARAM_DESCRIPTION, Default: names(fit)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
