% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkpar}
\alias{pkpar}
\title{FUNCTION_TITLE}
\usage{
pkpar(x, y)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
