% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qqhist}
\alias{qqhist}
\title{FUNCTION_TITLE}
\usage{
qqhist(x, label = NULL, main = NULL)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{label}{PARAM_DESCRIPTION, Default: NULL}

\item{main}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lattice]{histogram}},\code{\link[lattice]{qqmath}},\code{\link[lattice]{panel.qqmath}},\code{\link[lattice]{panel.qqmathline}}
}
