% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbeta2}
\alias{rbeta2}
\title{FUNCTION_TITLE}
\usage{
rbeta2(n, mean, sd)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}

\item{mean}{PARAM_DESCRIPTION}

\item{sd}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{rbeta}}
}
