% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbeta2Scaled}
\alias{rbeta2Scaled}
\title{FUNCTION_TITLE}
\usage{
rbeta2Scaled(n, mean, sd, lower, upper)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}

\item{mean}{PARAM_DESCRIPTION}

\item{sd}{PARAM_DESCRIPTION}

\item{lower}{PARAM_DESCRIPTION}

\item{upper}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
