% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rnorm.trunc}
\alias{rnorm.trunc}
\title{FUNCTION_TITLE}
\usage{
rnorm.trunc(n, mean = 0, sd = 1, lower = -Inf, upper = Inf)
}
\arguments{
\item{n}{PARAM_DESCRIPTION}

\item{mean}{PARAM_DESCRIPTION, Default: 0}

\item{sd}{PARAM_DESCRIPTION, Default: 1}

\item{lower}{PARAM_DESCRIPTION, Default: -Inf}

\item{upper}{PARAM_DESCRIPTION, Default: Inf}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
