% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDiagnose.R
\name{runDiagnose}
\alias{runDiagnose}
\title{FUNCTION_TITLE}
\usage{
runDiagnose(model, data, init, seed, chain = 1, refresh = 100)
}
\arguments{
\item{model}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{init}{PARAM_DESCRIPTION}

\item{seed}{PARAM_DESCRIPTION}

\item{chain}{PARAM_DESCRIPTION, Default: 1}

\item{refresh}{PARAM_DESCRIPTION, Default: 100}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
