% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModelFixed.R
\name{runModelFixed}
\alias{runModelFixed}
\title{FUNCTION_TITLE}
\usage{
runModelFixed(model, data, iter, warmup, thin, init, seed, chain = 1,
  stepsize = 1, adapt_delta = 0.8, max_depth = 10, refresh = 100)
}
\arguments{
\item{model}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{iter}{PARAM_DESCRIPTION}

\item{warmup}{PARAM_DESCRIPTION}

\item{thin}{PARAM_DESCRIPTION}

\item{init}{PARAM_DESCRIPTION}

\item{seed}{PARAM_DESCRIPTION}

\item{chain}{PARAM_DESCRIPTION, Default: 1}

\item{stepsize}{PARAM_DESCRIPTION, Default: 1}

\item{adapt_delta}{PARAM_DESCRIPTION, Default: 0.8}

\item{max_depth}{PARAM_DESCRIPTION, Default: 10}

\item{refresh}{PARAM_DESCRIPTION, Default: 100}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
