% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mcmc.list.R
\name{summary.mcmc.list}
\alias{summary.mcmc.list}
\title{FUNCTION_TITLE}
\usage{
\method{summary}{mcmc.list}(object, quantiles = c(0.025, 0.25, 0.5, 0.75,
  0.975), ...)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{quantiles}{PARAM_DESCRIPTION, Default: c(0.025, 0.25, 0.5, 0.75, 0.975)}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
