% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{swap}
\alias{swap}
\title{FUNCTION_TITLE}
\usage{
swap(x, oldval, newval)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{oldval}{PARAM_DESCRIPTION}

\item{newval}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
