% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{symmat}
\alias{symmat}
\title{FUNCTION_TITLE}
\usage{
symmat(a)
}
\arguments{
\item{a}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
